Fixes microsoft/monaco-editor#2650: Move scrollbar style participant down to the editor layer
